put everything into a folder
